Filter by gateway

+++ src/Http/Controllers/SocialGatewayController.php
@@ -26,7 +26,7 @@ class SocialGatewayController extends Controller
             }
 
             if (request()->has('gateway') && request()->input('gateway')) {
-                $condition['gateway'] = (int) request()->input('gateway');
+                $condition['gateway'] = request()->input('gateway');
             }
 
             if (request()->has('get_all') && (int) request()->input('get_all') === 1) {
